Refactor: organize reusable components, extract partials
