+++ src/components/InfoWidget/partials/ErrorState.tsx
@@ -1,0 +1,3 @@
+const ErrorState = () => <span>Unable to display weather information.</span>;
+
+export default ErrorState;
