+++ src/components/Sidebar/SidebarExpandButton.tsx
@@ -3,7 +3,7 @@ import { ChevronRightIcon } from "@heroicons/react/24/solid";
 import { cn } from "@/src/utils";
 
 import { Props } from "./Sidebar";
-import IconButton from "../IconButton";
+import IconButton from "../ui/IconButton";
 
 export const SidebarExpandButton = ({ isExpanded, setIsExpanded }: Props) => (
   <IconButton
